+++ api/src/main/java/com/programmers/bucketback/domains/vote/api/dto/response/VoteGetResponse.java
@@ -3,7 +3,7 @@ package com.programmers.bucketback.domains.vote.api.dto.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.programmers.bucketback.domains.item.model.ItemInfo;
 import com.programmers.bucketback.domains.vote.application.dto.response.VoteGetServiceResponse;
-import com.programmers.bucketback.domains.vote.model.response.VoteInfo;
+import com.programmers.bucketback.domains.vote.model.VoteInfo;
 
 import lombok.Builder;
