+++ domain/src/main/java/com/programmers/bucketback/domains/vote/implementation/VoteReader.java
@@ -16,10 +16,10 @@ import com.programmers.bucketback.domains.item.model.ItemInfo;
 import com.programmers.bucketback.domains.vote.domain.Vote;
 import com.programmers.bucketback.domains.vote.domain.Voter;
 import com.programmers.bucketback.domains.vote.model.VoteCursorSummary;
+import com.programmers.bucketback.domains.vote.model.VoteInfo;
+import com.programmers.bucketback.domains.vote.model.VoteSortCondition;
+import com.programmers.bucketback.domains.vote.model.VoteStatusCondition;
 import com.programmers.bucketback.domains.vote.model.VoteSummary;
-import com.programmers.bucketback.domains.vote.model.request.VoteSortCondition;
-import com.programmers.bucketback.domains.vote.model.request.VoteStatusCondition;
-import com.programmers.bucketback.domains.vote.model.response.VoteInfo;
 import com.programmers.bucketback.domains.vote.repository.VoteRepository;
 import com.programmers.bucketback.error.exception.BusinessException;
 import com.programmers.bucketback.error.exception.EntityNotFoundException;
