+++ domain/src/main/java/com/programmers/bucketback/domains/vote/model/VoteCursorSummary.java
@@ -1,7 +1,5 @@
 package com.programmers.bucketback.domains.vote.model;
 
-import com.programmers.bucketback.domains.vote.model.response.VoteInfo;
-
 public record VoteCursorSummary(
 	VoteInfo voteInfo,
 	Long item1Id,
