+++ domain/src/main/java/com/programmers/bucketback/domains/vote/model/VoteSummary.java
@@ -3,7 +3,6 @@ package com.programmers.bucketback.domains.vote.model;
 import com.programmers.bucketback.common.cursor.CursorIdParser;
 import com.programmers.bucketback.domains.item.domain.Item;
 import com.programmers.bucketback.domains.item.model.ItemInfo;
-import com.programmers.bucketback.domains.vote.model.response.VoteInfo;
 
 import lombok.Builder;
