+++ domain/src/main/java/com/programmers/bucketback/domains/vote/repository/VoteRepositoryForCursorImpl.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 import com.programmers.bucketback.Hobby;
 import com.programmers.bucketback.domains.vote.model.VoteCursorSummary;
-import com.programmers.bucketback.domains.vote.model.request.VoteSortCondition;
-import com.programmers.bucketback.domains.vote.model.request.VoteStatusCondition;
-import com.programmers.bucketback.domains.vote.model.response.VoteInfo;
+import com.programmers.bucketback.domains.vote.model.VoteInfo;
+import com.programmers.bucketback.domains.vote.model.VoteSortCondition;
+import com.programmers.bucketback.domains.vote.model.VoteStatusCondition;
 import com.querydsl.core.types.ConstantImpl;
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
